backend/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
python-dotenv==1.0.1
supabase==2.10.0
pydantic==2.10.3
python-multipart==0.0.20
cors==1.0.1